+++ lib/datamappify/repository/query_method/where.rb
@@ -4,9 +4,15 @@ module Datamappify
       class Where < Method
         # @return [Array<Entity>]
         def perform
-          dispatch_criteria_to_default_source(
+          results = dispatch_criteria_to_default_source(
             :Where, data_mapper.entity_class, criteria, data_mapper.attributes
           )
+
+          results.each do |entity|
+            states.find(entity)
+          end
+
+          results
         end
 
         # @see Method#reader?