+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 
 - Added preliminary support for composed search criteria.
+- Fixed dirty tracking for entities returned from `where` and `all`.
 
 ## 0.60.0 [2013-07-23]